feat: add a generic per-device key store
This change adds a couple very generic classes which handle creating and reading uniquely-identifiable files.

A `ValueBox` is a "container", which may correspond to a module or package, and has a 1:1 relationship with a particular directory on disk.
A `ValueWrapper` corresponds to a file in this container _and_ the contents of the file.  The file contents can be of any encoding supported by Node.js' `Buffer` (including `null`, which is just a `Buffer`).